folder restructure of attachment handlers, attachmentpreview dialog boxes and ImageGallery
