+++ packages/react/src/components/AttachmentHandler/PinnedAttachment.js
