+++ packages/react/src/components/AttachmentHandler/VideoAttachment.js
