+++ packages/react/src/components/AttachmentPreview/Backdrop.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Box } from '../../Box';
+import { Box } from '../Box';
 
 const Backdrop = ({ children, onClick }) => (
   <Box
