+++ packages/react/src/components/AttachmentPreview/preview/audio.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import { Box } from '../../../Box';
+import { Box } from '../../Box';
 
 function PreviewAudio({ previewURL }) {
   return (
