+++ packages/react/src/components/ImageGallery/ImageGallery.js
