+++ packages/react/src/components/ImageGallery/Swiper.js
